made runners run from 0

+++ src_java/FibonacciApproach.java
@@ -1,7 +1,7 @@
 public interface FibonacciApproach {
 
     //our algo function!
-    int fibonacci(int n);
+    long fibonacci(int n);
     
     //getter and setter for ops
     int getOps();
